- Change sound board path to /sound-board - Move sound board component to components folder - give grid layout
